comm message integration

+++ apps/adcsMC_app/fsw/inc/adcsMC_app_extern_typedefs.h
@@ -12,10 +12,10 @@
 
  typedef struct
  {
-     double quaternion[4];  /**< Quaternion representing the attitude */
-     double angular_velocity[3];  /**< Angular velocity vector */
-     double speed_vector[3];  /**< Speed vector */
-     double position_vector[3];  /**< Position vector */
+     float quaternion[4];  /**< Quaternion representing the attitude */
+     float angular_velocity[3];  /**< Angular velocity vector */
+     float speed_vector[3];  /**< Speed vector */
+     float position_vector[3];  /**< Position vector */
  } AdcsMC_MinimalTelemetry_t;
  
  #endif /* ADCSMC_EXTERN_TYPEDEFS_H */
